Separated server for auth and movies
